pyproject.toml
[project]
name = "qb-rule-editor"
version = "1.0.0"
description = "qBittorrent RSS规则编辑器"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
qb-rule-editor = "qb_rule_editor:main"

[tool.uv.sources]
# 使用uv的默认源配置